Remove fs-extra. Add export of "plugin". Rename to extendWithFirebaseConfig
