+++ docs.md
@@ -1,0 +1,166 @@
+<!-- Generated by documentation.js. Update this documentation by updating the source code. -->
+
+### Table of Contents
+
+-   [cy.login][1]
+    -   [Examples][2]
+-   [currentUser][3]
+-   [cy.logout][4]
+    -   [Examples][5]
+-   [cy.callRtdb][6]
+    -   [Parameters][7]
+    -   [Examples][8]
+-   [cy.callFirestore][9]
+    -   [Parameters][10]
+    -   [Examples][11]
+
+## cy.login
+
+Login to Firebase auth using FIREBASE_AUTH_JWT environment variable
+which is generated using firebase-admin authenticated with serviceAccount
+during test:buildConfig phase.
+
+Type: Cypress.Command
+
+### Examples
+
+```javascript
+cy.login()
+```
+
+## currentUser
+
+Log in using token \*
+
+## cy.logout
+
+Log out of Firebase instance
+
+Type: Cypress.Command
+
+### Examples
+
+```javascript
+cy.logout()
+```
+
+## cy.callRtdb
+
+Call Real Time Database path with some specified action. Authentication is through FIREBASE_TOKEN since firebase-tools is used (instead of firebaseExtra).
+
+Type: Cypress.Command
+
+### Parameters
+
+-   `action` **[String][12]** The action type to call with (set, push, update, remove)
+-   `actionPath` **[String][12]** Path within RTDB that action should be applied
+-   `opts` **[Object][13]** Options
+    -   `opts.args` **[Array][14]** Command line args to be passed
+
+### Examples
+
+Set Data
+
+
+```javascript
+const fakeProject = { some: 'data' }
+cy.callRtdb('set', 'projects/ABC123', fakeProject)
+```
+
+Set Data With Meta
+
+
+```javascript
+const fakeProject = { some: 'data' }
+// Adds createdAt and createdBy (current user's uid) on data
+cy.callRtdb('set', 'projects/ABC123', fakeProject, { withMeta: true })
+```
+
+Get/Verify Data
+
+
+```javascript
+cy.callRtdb('get', 'projects/ABC123')
+  .then((project) => {
+    // Confirm new data has users uid
+    cy.wrap(project)
+      .its('createdBy')
+      .should('equal', Cypress.env('TEST_UID'))
+  })
+```
+
+Other Args
+
+
+```javascript
+const opts = { args: ['-d'] }
+const fakeProject = { some: 'data' }
+cy.callRtdb('update', 'project/test-project', fakeProject, opts)
+```
+
+## cy.callFirestore
+
+Call Firestore instance with some specified action. Authentication is through serviceAccount.json since it is at the base
+level. If using delete, auth is through `FIREBASE_TOKEN` since firebase-tools is used (instead of firebaseExtra).
+
+Type: Cypress.Command
+
+### Parameters
+
+-   `action` **[String][12]** The action type to call with (set, push, update, remove)
+-   `actionPath` **[String][12]** Path within RTDB that action should be applied
+-   `opts` **[Object][13]** Options
+    -   `opts.args` **[Array][14]** Command line args to be passed
+
+### Examples
+
+Basic
+
+
+```javascript
+cy.callFirestore('add', 'project/test-project', 'fakeProject.json')
+```
+
+Recursive Delete
+
+
+```javascript
+const opts = { recursive: true }
+cy.callFirestore('delete', 'project/test-project', opts)
+```
+
+Other Args
+
+
+```javascript
+const opts = { args: ['-r'] }
+cy.callFirestore('delete', 'project/test-project', opts)
+```
+
+[1]: #cylogin
+
+[2]: #examples
+
+[3]: #currentuser
+
+[4]: #cylogout
+
+[5]: #examples-1
+
+[6]: #cycallrtdb
+
+[7]: #parameters
+
+[8]: #examples-2
+
+[9]: #cycallfirestore
+
+[10]: #parameters-1
+
+[11]: #examples-3
+
+[12]: https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/String
+
+[13]: https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/Object
+
+[14]: https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/Array
