+++ package.json
@@ -23,8 +23,8 @@
   "dependencies": {
     "chalk": "^2.4.1",
     "commander": "^2.19.0",
+    "figures": "^2.0.0",
     "firebase-admin": "^6.0.0",
-    "fs-extra": "^7.0.1",
     "lodash": "^4.17.11"
   },
   "devDependencies": {
